Refactor everything in samples route tests into the MockDataFactory
