+++ src/api/route-services/__mocks__/MockDataFactory.js
@@ -1,3 +1,190 @@
+const singleSampleProcessingConfig = (sampleId) => ({
+  cellSizeDistribution: {
+    [sampleId]: {
+      auto: true,
+      filterSettings: {
+        minCellSize: 516,
+        binStep: 200,
+      },
+      defaultFilterSettings: {
+        minCellSize: 516,
+        binStep: 200,
+      },
+    },
+    enabled: false,
+  },
+  classifier: {
+    [sampleId]: {
+      auto: true,
+      filterSettings: {
+        FDR: 0.01,
+      },
+      prefiltered: true,
+      defaultFilterSettings: {
+        FDR: 0.01,
+      },
+    },
+    prefiltered: true,
+    filterSettings: {
+      FDR: 0.01,
+    },
+    enabled: false,
+  },
+  dataIntegration: {
+    dataIntegration: {
+      method: 'harmony',
+      methodSettings: {
+        harmony: {
+          numGenes: 2000,
+          normalisation: 'logNormalize',
+        },
+        fastmnn: {
+          numGenes: 2000,
+          normalisation: 'logNormalize',
+        },
+        seuratv4: {
+          numGenes: 2000,
+          normalisation: 'logNormalize',
+        },
+        unisample: {
+          numGenes: 2000,
+          normalisation: 'logNormalize',
+        },
+      },
+    },
+    dimensionalityReduction: {
+      method: 'rpca',
+      excludeGeneCategories: [],
+      numPCs: 30,
+    },
+  },
+  mitochondrialContent: {
+    [sampleId]: {
+      auto: true,
+      filterSettings: {
+        method: 'absolute_threshold',
+        methodSettings: {
+          absolute_threshold: {
+            maxFraction: 0.1,
+            binStep: 0.05,
+          },
+        },
+      },
+      defaultFilterSettings: {
+        method: 'absolute_threshold',
+        methodSettings: {
+          absolute_threshold: {
+            maxFraction: 0.1,
+            binStep: 0.05,
+          },
+        },
+      },
+    },
+    enabled: true,
+  },
+  configureEmbedding: {
+    embeddingSettings: {
+      method: 'umap',
+      methodSettings: {
+        tsne: {
+          perplexity: 9.18,
+          learningRate: 200,
+        },
+        umap: {
+          minimumDistance: 0.3,
+          distanceMetric: 'cosine',
+        },
+      },
+    },
+    clusteringSettings: {
+      method: 'louvain',
+      methodSettings: {
+        louvain: {
+          resolution: 0.8,
+        },
+      },
+    },
+  },
+  numGenesVsNumUmis: {
+    [sampleId]: {
+      auto: true,
+      filterSettings: {
+        regressionType: 'gam',
+        regressionTypeSettings: {
+          gam: {
+            'p.level': 0.001,
+          },
+        },
+      },
+      defaultFilterSettings: {
+        regressionType: 'gam',
+        regressionTypeSettings: {
+          gam: {
+            'p.level': 0.001,
+          },
+        },
+      },
+    },
+    enabled: true,
+  },
+  doubletScores: {
+    auto: true,
+    [sampleId]: {
+      auto: true,
+      auth_JWT: '',
+      filterSettings: {
+        probabilityThreshold: 0.7622447,
+        binStep: 0.05,
+      },
+      defaultFilterSettings: {
+        probabilityThreshold: 0.7622447,
+        binStep: 0.05,
+      },
+    },
+    enabled: true,
+  },
+});
+
+const filesObject = {
+  lastModified: '2021-10-22T12:39:41.117Z',
+  'matrix.mtx.gz': {
+    valid: true,
+    path: '9b5d23ef-b447-4239-8e17-714aa9a4d06e/matrix.mtx.gz',
+    upload: {},
+    name: 'matrix.mtx.gz',
+    compressed: true,
+    lastModified: '2021-10-22T12:39:41.033Z',
+    bundle: {
+      path: '/WT1/matrix.mtx.gz',
+    },
+    errors: '',
+  },
+  'features.tsv.gz': {
+    valid: true,
+    path: '9b5d23ef-b447-4239-8e17-714aa9a4d06e/matrix.mtx.gz',
+    upload: {},
+    name: 'matrix.mtx.gz',
+    compressed: true,
+    lastModified: '2021-10-22T12:39:41.033Z',
+    bundle: {
+      path: '/WT1/matrix.mtx.gz',
+    },
+    errors: '',
+  },
+  'barcodes.tsv.gz': {
+    valid: true,
+    path: '9b5d23ef-b447-4239-8e17-714aa9a4d06e/matrix.mtx.gz',
+    upload: {},
+    name: 'matrix.mtx.gz',
+    compressed: true,
+    lastModified: '2021-10-22T12:39:41.033Z',
+    bundle: {
+      path: '/WT1/matrix.mtx.gz',
+    },
+    errors: '',
+  },
+};
+
 class MockDataFactory {
   constructor(values = {}) {
     const {
@@ -10,152 +197,8 @@ class MockDataFactory {
     this.projectId = projectId;
     this.sampleIds = [sampleId];
 
-    this.processingConfig = {
-      cellSizeDistribution: {
-        [sampleId]: {
-          auto: true,
-          filterSettings: {
-            minCellSize: 516,
-            binStep: 200,
-          },
-          defaultFilterSettings: {
-            minCellSize: 516,
-            binStep: 200,
-          },
-        },
-        enabled: false,
-      },
-      classifier: {
-        [sampleId]: {
-          auto: true,
-          filterSettings: {
-            FDR: 0.01,
-          },
-          prefiltered: true,
-          defaultFilterSettings: {
-            FDR: 0.01,
-          },
-        },
-        prefiltered: true,
-        filterSettings: {
-          FDR: 0.01,
-        },
-        enabled: false,
-      },
-      dataIntegration: {
-        dataIntegration: {
-          method: 'harmony',
-          methodSettings: {
-            harmony: {
-              numGenes: 2000,
-              normalisation: 'logNormalize',
-            },
-            fastmnn: {
-              numGenes: 2000,
-              normalisation: 'logNormalize',
-            },
-            seuratv4: {
-              numGenes: 2000,
-              normalisation: 'logNormalize',
-            },
-            unisample: {
-              numGenes: 2000,
-              normalisation: 'logNormalize',
-            },
-          },
-        },
-        dimensionalityReduction: {
-          method: 'rpca',
-          excludeGeneCategories: [],
-          numPCs: 30,
-        },
-      },
-      mitochondrialContent: {
-        [sampleId]: {
-          auto: true,
-          filterSettings: {
-            method: 'absolute_threshold',
-            methodSettings: {
-              absolute_threshold: {
-                maxFraction: 0.1,
-                binStep: 0.05,
-              },
-            },
-          },
-          defaultFilterSettings: {
-            method: 'absolute_threshold',
-            methodSettings: {
-              absolute_threshold: {
-                maxFraction: 0.1,
-                binStep: 0.05,
-              },
-            },
-          },
-        },
-        enabled: true,
-      },
-      configureEmbedding: {
-        embeddingSettings: {
-          method: 'umap',
-          methodSettings: {
-            tsne: {
-              perplexity: 9.18,
-              learningRate: 200,
-            },
-            umap: {
-              minimumDistance: 0.3,
-              distanceMetric: 'cosine',
-            },
-          },
-        },
-        clusteringSettings: {
-          method: 'louvain',
-          methodSettings: {
-            louvain: {
-              resolution: 0.8,
-            },
-          },
-        },
-      },
-      numGenesVsNumUmis: {
-        [sampleId]: {
-          auto: true,
-          filterSettings: {
-            regressionType: 'gam',
-            regressionTypeSettings: {
-              gam: {
-                'p.level': 0.001,
-              },
-            },
-          },
-          defaultFilterSettings: {
-            regressionType: 'gam',
-            regressionTypeSettings: {
-              gam: {
-                'p.level': 0.001,
-              },
-            },
-          },
-        },
-        enabled: true,
-      },
-      doubletScores: {
-        auto: true,
-        [sampleId]: {
-          auto: true,
-          auth_JWT: '',
-          filterSettings: {
-            probabilityThreshold: 0.7622447,
-            binStep: 0.05,
-          },
-          defaultFilterSettings: {
-            probabilityThreshold: 0.7622447,
-            binStep: 0.05,
-          },
-        },
-        enabled: true,
-      },
-    };
+    this.filesObject = filesObject;
+    this.processingConfig = singleSampleProcessingConfig(sampleId);
   }
 
   getExperiment() {
@@ -197,6 +240,27 @@ class MockDataFactory {
       uuid: this.projectId,
       samples: this.sampleIds,
     };
+  }
+
+  getSamplesEntry() {
+    return this.sampleIds.reduce((acum, sampleId) => {
+      // eslint-disable-next-line no-param-reassign
+      acum[sampleId] = {
+        name: 'sample-1',
+        projectUuid: 'project-1',
+        uuid: 'sample-1',
+        type: '10X Chromium',
+        species: 'hsapies',
+        createdDate: '2020-01-01T00:00:00.000Z',
+        lastModified: null,
+        complete: true,
+        error: false,
+        fileNames: ['test-1'],
+        files: this.filesObject,
+      };
+
+      return acum;
+    }, {});
   }
 }
